requirements.txt
openenv-core>=0.1.13
numpy
pydantic